refactor: move spanner::Date to common library
We are going to need this class in google/cloud/storage, the first step
is to refactor it.
