+++ google/cloud/spanner/date.h
@@ -16,9 +16,7 @@
 #define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_SPANNER_DATE_H
 
 #include "google/cloud/spanner/version.h"
-#include <cstdint>
-#include <ostream>
-#include <tuple>
+#include "google/cloud/date.h"
 
 namespace google {
 namespace cloud {
@@ -32,40 +30,7 @@ inline namespace SPANNER_CLIENT_NS {
  * Standard C++11 lacks a good "date" abstraction, so we supply a simple
  * implementation of a YMD triple.
  */
-class Date {
- public:
-  Date(std::int64_t year, int month, int day);  // Fields will be normalized.
-  Date() : Date(1970, 1, 1) {}
-
-  std::int64_t year() const { return year_; }
-  int month() const { return month_; }
-  int day() const { return day_; }
-
- private:
-  std::int64_t year_;
-  int month_;
-  int day_;
-};
-
-std::ostream& operator<<(std::ostream& os, Date const& date);
-
-inline bool operator==(Date const& a, Date const& b) {
-  return std::make_tuple(a.year(), a.month(), a.day()) ==
-         std::make_tuple(b.year(), b.month(), b.day());
-}
-
-inline bool operator!=(Date const& a, Date const& b) { return !(a == b); }
-
-inline bool operator<(Date const& a, Date const& b) {
-  return std::make_tuple(a.year(), a.month(), a.day()) <
-         std::make_tuple(b.year(), b.month(), b.day());
-}
-
-inline bool operator<=(Date const& a, Date const& b) { return !(b < a); }
-
-inline bool operator>=(Date const& a, Date const& b) { return !(a < b); }
-
-inline bool operator>(Date const& a, Date const& b) { return b < a; }
+using Date = ::google::cloud::Date;
 
 }  // namespace SPANNER_CLIENT_NS
 }  // namespace spanner
