+++ google/cloud/spanner/spanner_client.bzl
@@ -43,7 +43,6 @@ spanner_client_hdrs = [
     "internal/database_admin_logging.h",
     "internal/database_admin_metadata.h",
     "internal/database_admin_stub.h",
-    "internal/date.h",
     "internal/instance_admin_logging.h",
     "internal/instance_admin_metadata.h",
     "internal/instance_admin_stub.h",
@@ -96,7 +95,6 @@ spanner_client_srcs = [
     "database.cc",
     "database_admin_client.cc",
     "database_admin_connection.cc",
-    "date.cc",
     "instance.cc",
     "instance_admin_client.cc",
     "instance_admin_connection.cc",
@@ -106,7 +104,6 @@ spanner_client_srcs = [
     "internal/database_admin_logging.cc",
     "internal/database_admin_metadata.cc",
     "internal/database_admin_stub.cc",
-    "internal/date.cc",
     "internal/instance_admin_logging.cc",
     "internal/instance_admin_metadata.cc",
     "internal/instance_admin_stub.cc",
