+++ google/cloud/spanner/spanner_client_unit_tests.bzl
@@ -26,7 +26,6 @@ spanner_client_unit_tests = [
     "database_admin_client_test.cc",
     "database_admin_connection_test.cc",
     "database_test.cc",
-    "date_test.cc",
     "instance_admin_client_test.cc",
     "instance_admin_connection_test.cc",
     "instance_test.cc",
@@ -36,7 +35,6 @@ spanner_client_unit_tests = [
     "internal/connection_impl_test.cc",
     "internal/database_admin_logging_test.cc",
     "internal/database_admin_metadata_test.cc",
-    "internal/date_test.cc",
     "internal/instance_admin_logging_test.cc",
     "internal/instance_admin_metadata_test.cc",
     "internal/log_wrapper_test.cc",
